added an attribute to control inactive account deletion

+++ cookbooks/prepare/recipes/mac.rb
@@ -88,11 +88,13 @@ cookbook_file "/System/Library/User Template/English.lproj/Library/Preferences/c
 ### Deletes inactive users
 ruby_block "Delete inactive users" do
   block do
-    days = 14 # number of days until a user account is removed from the system
-    Dir.entries("/Users").each do |user|
-      if not (user =~ /^\./) and user != "Shared" and user != "administrator" and user != "labuser" and File.atime("/Users/#{user}") < (Time.now - days*(60*60*24))
-        system("dscl  . delete /Users/#{user}")
-        system("rm -rf /Users/#{user}")
+    days = node["prepare"]["inactivity"] # number of days until a user account is removed from the system
+    if days != 0
+      Dir.entries("/Users").each do |user|
+        if not (user =~ /^\./) and user != "Shared" and user != "administrator" and user != "labuser" and File.atime("/Users/#{user}") < (Time.now - days*(60*60*24))
+          system("dscl  . delete /Users/#{user}")
+          system("rm -rf /Users/#{user}")
+        end
       end
     end
   end
